refactor(adapters): extract remaining section renderers and tests to section.rs [#305-7/7]
Move render_license_compliance and render_resolution_guide from mod.rs
into section.rs as pub(super) free functions. Move 9 resolution guide
unit tests to section.rs. Inline section:: and vuln_render:: calls
directly in SbomFormatter::format, eliminating all wrapper methods.
mod.rs production code reduced to ~93 lines (struct, constructors,
Default, SbomFormatter::format only).

Closes #329
